Begin splitting up duckdb_python.cpp into separate files
