+++ tools/pythonpkg/MANIFEST.in
@@ -2,3 +2,4 @@ recursive-include tests *.py
 recursive-include duckdb_query_graph *.js *.css
 recursive-include duckdb *.h *.hpp *.cpp
 include duckdb_python.cpp
+recursive-include src *.h *.hpp *.cpp
